refactor(bom): move the subpart graph and QoH recalc out of the movement trigger
`loadSubpartGraph`, `getDeductionTargets` and `batchRecalculateQoH` were private
to `field-hooks/post/bom-movement-triggers.ts`. The build event needs the first
and the third, and a trigger module is the wrong place to reach into for them.

Pure move — the four bodies (including `deriveStockStatus`, which had no other
caller and so had to travel with `batchRecalculateQoH`) are byte-for-byte
identical to their previous versions. The trigger file gains exactly two lines,
both imports, and loses 382. No existing test was modified.

Adds `loadDirectSubparts(db, organizationId, partInstanceId)` — the depth-1
case as a plain two-join query, no recursion.

That is a separate function rather than a depth argument on the existing walk
because the distinction is load-bearing (build plan B4). `getDeductionTargets`
deducts every descendant *including intermediates*, which is right for
backflush-at-sale and wrong the moment a subassembly has its own on-hand
balance — which is exactly what build-to-stock creates. A build consumes its
part's DIRECT subparts only; a subassembly is produced by its own build. A
depth flag is a thing someone passes wrong; two named functions are not.

They live side by side in `subpart-graph.ts` for the same reason: the property
only holds if a reader meets both at once, so `loadDirectSubparts` can name in
its doc comment the walk it is deliberately not.

Not merged with `cost-calculator.ts`'s private `buildSubpartGraph`, despite the
matching return shape: that one is a pure in-memory reducer over rows the caller
already loaded org-wide, this one is a recursive CTE scoped to a single root.

⚠️ `loadSubpartGraph` and `batchRecalculateQoH` still close over the
module-level `database` singleton, so a caller cannot hand them a transaction —
only `loadDirectSubparts` takes `db` first. Fine for the build event as planned,
whose QoH recalc must run after commit regardless, but converting the two
legacy signatures is a follow-up.

Refs plans/products/build/01-build-plan.md §3.2
