+++ packages/lib/src/bom/index.ts
@@ -1,6 +1,8 @@
 // packages/lib/src/bom/index.ts
 
 export { recalculateAffectedParts, recalculateAllPartCosts } from './cost-calculator'
+export { batchRecalculateQoH } from './qoh'
+export { getDeductionTargets, loadDirectSubparts, loadSubpartGraph } from './subpart-graph'
 export type { LandedCostBreakdown, VendorCostRow } from './vendor-cost'
 export {
   computeLandedBreakdown,
