+++ packages/lib/src/field-hooks/post/bom-movement-triggers.ts
@@ -3,25 +3,18 @@
 import { database, schema } from '@auxx/database'
 import { createScopedLogger } from '@auxx/logger'
 import type { TypedFieldValueInput } from '@auxx/types'
-import { buildFieldValueKey, type FieldId } from '@auxx/types/field'
 import type { RecordId } from '@auxx/types/resource'
 import { parseRecordId, toRecordId } from '@auxx/types/resource'
 import { nextKeyAfter } from '@auxx/utils/fractional-indexing'
-import { and, eq, inArray, sql } from 'drizzle-orm'
+import { batchRecalculateQoH } from '../../bom/qoh'
+import { getDeductionTargets, loadSubpartGraph } from '../../bom/subpart-graph'
 import { getOrgCache, requireCachedEntityDefId } from '../../cache'
 import { createFieldValueContext } from '../../field-values/field-value-helpers'
 import { buildFieldValueRow, setValueWithType } from '../../field-values/field-value-mutations'
 import { type StoredFieldType, toFieldType } from '../../field-values/stored-field-type'
-import {
-  type FieldValueUpdateEntry,
-  getRealtimeService,
-  publishFieldValueUpdates,
-} from '../../realtime'
 import type { EntityTriggerHandler } from '../types'
 
 const logger = createScopedLogger('field-hooks:bom-movement')
-
-const MAX_BOM_DEPTH = 20
 
 // ─── Main Trigger Handler ──────────────────────────────────────────────
 
@@ -209,372 +202,6 @@ export const explodeBomMovement: EntityTriggerHandler = async (event) => {
   })
 }
 
-// ─── Subpart Graph Loader ──────────────────────────────────────────────
-
-/**
- * Load the subpart graph for a specific part's subtree using a recursive CTE.
- * Returns only the parent→child edges reachable from rootPartId.
- *
- * The CTE has a depth limit (MAX_BOM_DEPTH) which:
- * 1. Scopes the query to only relevant relationships (not entire org)
- * 2. Prevents infinite recursion from circular BOM refs at the DB level
- * 3. Acts as a safety valve for unreasonably deep BOMs
- */
-async function loadSubpartGraph(
-  organizationId: string,
-  rootPartId: string
-): Promise<Map<string, { childId: string; qty: number }[]>> {
-  const cache = getOrgCache()
-  const subpartDefId = await requireCachedEntityDefId(organizationId, 'subpart')
-
-  const cfFields = await cache
-    .from(organizationId, 'customFields')
-    .bySystemAttributes(['subpart_parent_part', 'subpart_child_part', 'subpart_quantity'] as const)
-
-  const spParentField = cfFields.subpart_parent_part
-  const spChildField = cfFields.subpart_child_part
-  const spQtyField = cfFields.subpart_quantity
-
-  if (!spParentField || !spChildField || !spQtyField) return new Map()
-
-  // Recursive CTE: walk from rootPartId down through the subpart graph.
-  // Each row returned is a (parentId, childId, qty) edge in the subtree.
-  const result = await database.execute(sql`
-    WITH RECURSIVE subtree AS (
-      -- Base case: direct children of the root part
-      SELECT
-        parent_fv."relatedEntityId" AS parent_id,
-        child_fv."relatedEntityId" AS child_id,
-        qty_fv."valueNumber" AS qty,
-        1 AS depth
-      FROM ${schema.EntityInstance} ei
-      JOIN ${schema.FieldValue} parent_fv
-        ON parent_fv."entityId" = ei.id
-        AND parent_fv."fieldId" = ${spParentField.id}
-        AND parent_fv."organizationId" = ${organizationId}
-      JOIN ${schema.FieldValue} child_fv
-        ON child_fv."entityId" = ei.id
-        AND child_fv."fieldId" = ${spChildField.id}
-        AND child_fv."organizationId" = ${organizationId}
-      JOIN ${schema.FieldValue} qty_fv
-        ON qty_fv."entityId" = ei.id
-        AND qty_fv."fieldId" = ${spQtyField.id}
-        AND qty_fv."organizationId" = ${organizationId}
-      WHERE ei."organizationId" = ${organizationId}
-        AND ei."entityDefinitionId" = ${subpartDefId}
-        AND ei."archivedAt" IS NULL
-        AND parent_fv."relatedEntityId" = ${rootPartId}
-
-      UNION ALL
-
-      -- Recursive step: follow each child to find its children
-      SELECT
-        parent_fv."relatedEntityId" AS parent_id,
-        child_fv."relatedEntityId" AS child_id,
-        qty_fv."valueNumber" AS qty,
-        st.depth + 1 AS depth
-      FROM subtree st
-      JOIN ${schema.FieldValue} parent_fv
-        ON parent_fv."relatedEntityId" = st.child_id
-        AND parent_fv."fieldId" = ${spParentField.id}
-        AND parent_fv."organizationId" = ${organizationId}
-      JOIN ${schema.EntityInstance} ei
-        ON ei.id = parent_fv."entityId"
-        AND ei."organizationId" = ${organizationId}
-        AND ei."entityDefinitionId" = ${subpartDefId}
-        AND ei."archivedAt" IS NULL
-      JOIN ${schema.FieldValue} child_fv
-        ON child_fv."entityId" = ei.id
-        AND child_fv."fieldId" = ${spChildField.id}
-        AND child_fv."organizationId" = ${organizationId}
-      JOIN ${schema.FieldValue} qty_fv
-        ON qty_fv."entityId" = ei.id
-        AND qty_fv."fieldId" = ${spQtyField.id}
-        AND qty_fv."organizationId" = ${organizationId}
-      WHERE st.depth < ${MAX_BOM_DEPTH}
-    )
-    SELECT parent_id, child_id, qty FROM subtree
-  `)
-
-  const rows = result.rows as { parent_id: string; child_id: string; qty: number }[]
-
-  // Build adjacency list from CTE results
-  const graph = new Map<string, { childId: string; qty: number }[]>()
-  for (const row of rows) {
-    if (row.parent_id && row.child_id && row.qty > 0) {
-      const children = graph.get(row.parent_id) ?? []
-      children.push({ childId: row.child_id, qty: row.qty })
-      graph.set(row.parent_id, children)
-    }
-  }
-
-  return graph
-}
-
-// ─── Leaf Target Calculation ───────────────────────────────────────────
-
-/**
- * Walk the BOM graph from root and collect all DESCENDANT parts with multiplied
- * quantities. The root itself is excluded — the user-submitted parent movement
- * already accounts for the root's deduction. Both intermediate and leaf
- * descendants are included (all have tracked inventory).
- * Includes circular reference detection via visited Set and maxDepth safeguard.
- */
-function getDeductionTargets(
-  rootPartId: string,
-  rootQuantity: number,
-  graph: Map<string, { childId: string; qty: number }[]>,
-  visited: Set<string> = new Set(),
-  depth: number = 0,
-  isRoot: boolean = true
-): { partInstanceId: string; quantity: number }[] {
-  // Circular reference protection — same pattern as cost-calculator.ts
-  if (visited.has(rootPartId)) {
-    logger.warn('Circular reference detected in BOM during stock explosion, skipping', {
-      partId: rootPartId,
-      depth,
-    })
-    return []
-  }
-
-  // Max depth safeguard — defense in depth alongside the CTE's depth limit
-  if (depth >= MAX_BOM_DEPTH) {
-    logger.warn('Max BOM depth reached during stock explosion, treating as leaf', {
-      partId: rootPartId,
-      depth,
-    })
-    return isRoot ? [] : [{ partInstanceId: rootPartId, quantity: rootQuantity }]
-  }
-
-  visited.add(rootPartId)
-
-  const children = graph.get(rootPartId)
-
-  // No children = leaf node. Skip if this is the root (parent movement covers it).
-  if (!children || children.length === 0) {
-    return isRoot ? [] : [{ partInstanceId: rootPartId, quantity: rootQuantity }]
-  }
-
-  // Has children = include this node (unless it's the root) AND recurse into children
-  const targets: { partInstanceId: string; quantity: number }[] = isRoot
-    ? []
-    : [{ partInstanceId: rootPartId, quantity: rootQuantity }]
-  for (const child of children) {
-    targets.push(
-      ...getDeductionTargets(
-        child.childId,
-        rootQuantity * child.qty,
-        graph,
-        visited,
-        depth + 1,
-        false
-      )
-    )
-  }
-
-  // Consolidate: a part may appear multiple times in the tree
-  const consolidated = new Map<string, number>()
-  for (const t of targets) {
-    consolidated.set(t.partInstanceId, (consolidated.get(t.partInstanceId) ?? 0) + t.quantity)
-  }
-
-  return [...consolidated.entries()].map(([partInstanceId, quantity]) => ({
-    partInstanceId,
-    quantity,
-  }))
-}
-
-// ─── Batch QoH Recalculation ───────────────────────────────────────────
-
-/**
- * Recalculate QoH for multiple parts in one pass.
- * - 1 grouped SUM query for all parts
- * - 1 batch reorder point read
- * - Parallel writes for QoH + stock_status
- * - 1 batched realtime publish
- */
-async function batchRecalculateQoH(
-  organizationId: string,
-  partInstanceIds: string[]
-): Promise<void> {
-  if (partInstanceIds.length === 0) return
-
-  const unique = [...new Set(partInstanceIds)]
-  const cache = getOrgCache()
-
-  const fields = await cache
-    .from(organizationId, 'customFields')
-    .bySystemAttributes([
-      'stock_movement_quantity',
-      'stock_movement_part',
-      'stock_movement_adjust_subparts',
-      'part_quantity_on_hand',
-      'part_reorder_point',
-      'part_stock_status',
-    ] as const)
-
-  const qtyField = fields.stock_movement_quantity
-  const partRelField = fields.stock_movement_part
-  const flagField = fields.stock_movement_adjust_subparts
-  const qohField = fields.part_quantity_on_hand
-  const reorderPointField = fields.part_reorder_point
-  const statusField = fields.part_stock_status
-
-  if (!qtyField || !partRelField || !qohField) return
-
-  // 1. Grouped SUM: one query for all parts
-  //    Excludes movements where adjust_subparts=true
-  const sumRows = await database
-    .select({
-      partId: sql<string>`fv_part."relatedEntityId"`,
-      total: sql<string>`COALESCE(SUM(${schema.FieldValue.valueNumber}), 0)`,
-    })
-    .from(schema.FieldValue)
-    .innerJoin(
-      sql`"FieldValue" fv_part`,
-      sql`${schema.FieldValue.entityId} = fv_part."entityId"
-        AND fv_part."fieldId" = ${partRelField.id}
-        AND fv_part."relatedEntityId" IN (${sql.join(
-          unique.map((id) => sql`${id}`),
-          sql`, `
-        )})
-        AND fv_part."organizationId" = ${organizationId}`
-    )
-    .leftJoin(
-      sql`"FieldValue" fv_flag`,
-      sql`${schema.FieldValue.entityId} = fv_flag."entityId"
-        AND fv_flag."fieldId" = ${flagField?.id ?? ''}
-        AND fv_flag."organizationId" = ${organizationId}`
-    )
-    .where(
-      and(
-        eq(schema.FieldValue.fieldId, qtyField.id),
-        eq(schema.FieldValue.organizationId, organizationId),
-        sql`(fv_flag."valueBoolean" IS NULL OR fv_flag."valueBoolean" = false)`
-      )
-    )
-    .groupBy(sql`fv_part."relatedEntityId"`)
-
-  const qohByPart = new Map<string, number>()
-  for (const row of sumRows) {
-    qohByPart.set(row.partId, Number(row.total ?? 0))
-  }
-  // Parts with zero movements won't appear — default to 0
-  for (const id of unique) {
-    if (!qohByPart.has(id)) qohByPart.set(id, 0)
-  }
-
-  // 2. Batch read reorder points (1 query)
-  const reorderPoints = new Map<string, number | null>()
-  if (reorderPointField) {
-    const rpRows = await database
-      .select({
-        entityId: schema.FieldValue.entityId,
-        valueNumber: schema.FieldValue.valueNumber,
-      })
-      .from(schema.FieldValue)
-      .where(
-        and(
-          inArray(schema.FieldValue.entityId, unique),
-          eq(schema.FieldValue.organizationId, organizationId),
-          eq(schema.FieldValue.fieldId, reorderPointField.id)
-        )
-      )
-    for (const row of rpRows) {
-      reorderPoints.set(row.entityId, row.valueNumber != null ? Number(row.valueNumber) : null)
-    }
-  }
-
-  // 3. Batch write QoH + stock_status (2 queries: 1 bulk delete + 1 bulk insert)
-  const partDefId = await requireCachedEntityDefId(organizationId, 'part')
-  const realtimeEntries: FieldValueUpdateEntry[] = []
-
-  const insertRows: Array<typeof schema.FieldValue.$inferInsert> = []
-
-  for (const partId of unique) {
-    const qoh = qohByPart.get(partId) ?? 0
-    const reorderPoint = reorderPoints.get(partId) ?? null
-    const status = deriveStockStatus(qoh, reorderPoint)
-    const recordId = toRecordId(partDefId, partId) as RecordId
-
-    // QoH field value row (single-value field; positional sortKey).
-    insertRows.push(
-      buildFieldValueRow({
-        organizationId,
-        entityId: partId,
-        entityDefinitionId: partDefId,
-        fieldId: qohField.id,
-        fieldType: toFieldType(qohField.type),
-        value: { type: 'number', value: qoh },
-        sortKey: nextKeyAfter(null),
-      })
-    )
-
-    // Stock status field value row
-    if (statusField) {
-      insertRows.push(
-        buildFieldValueRow({
-          organizationId,
-          entityId: partId,
-          entityDefinitionId: partDefId,
-          fieldId: statusField.id,
-          fieldType: toFieldType(statusField.type),
-          value: { type: 'option', optionId: status },
-          sortKey: nextKeyAfter(null),
-        })
-      )
-    }
-
-    // Collect realtime entries
-    realtimeEntries.push({
-      key: buildFieldValueKey(recordId, qohField.id as FieldId),
-      value: { type: 'number', value: qoh },
-    })
-    if (statusField) {
-      realtimeEntries.push({
-        key: buildFieldValueKey(recordId, statusField.id as FieldId),
-        value: { type: 'option', optionId: status },
-      })
-    }
-  }
-
-  // 1 bulk DELETE + 1 bulk INSERT, atomically
-  // (plans/field-values/delete-insert-replace.md Phase 0): a crash between
-  // the statements must not wipe stock levels for the whole batch.
-  // Cross-entity bulk replace, so no per-(entity, field) advisory lock —
-  // the transaction alone closes the destroy window.
-  const fieldIds = [qohField.id, ...(statusField ? [statusField.id] : [])]
-  await database.transaction(async (tx) => {
-    await tx
-      .delete(schema.FieldValue)
-      .where(
-        and(
-          inArray(schema.FieldValue.entityId, unique),
-          inArray(schema.FieldValue.fieldId, fieldIds),
-          eq(schema.FieldValue.organizationId, organizationId)
-        )
-      )
-
-    if (insertRows.length > 0) {
-      await tx.insert(schema.FieldValue).values(insertRows)
-    }
-  })
-
-  // 4. One batched realtime publish
-  if (realtimeEntries.length > 0) {
-    publishFieldValueUpdates(getRealtimeService(), organizationId, realtimeEntries).catch((err) => {
-      logger.error('Failed to publish batch QoH realtime update', {
-        error: err instanceof Error ? err.message : String(err),
-      })
-    })
-  }
-
-  logger.info('Batch QoH recalculation complete', {
-    parts: unique.length,
-    organizationId,
-  })
-}
-
 // ─── Helpers ───────────────────────────────────────────────────────────
 
 /**
@@ -603,13 +230,6 @@ async function clearAdjustSubpartsFlag(
     fieldType: toFieldType(flagField.type),
     value: { type: 'boolean', value: false },
   })
-}
-
-/** Derive stock status from QoH and reorder point. */
-function deriveStockStatus(qoh: number, reorderPoint: number | null): string {
-  if (qoh <= 0) return 'out_of_stock'
-  if (reorderPoint != null && qoh <= reorderPoint) return 'low_stock'
-  return 'in_stock'
 }
 
 /** Extract a related entity ID from event values. */
